refactor: Make extensions use GdrivePlayer lib
