+++ src/ar/anime4up/build.gradle
@@ -10,6 +10,7 @@ ext {
 }
 
 dependencies {
+    implementation(project(':lib-gdriveplayer-extractor'))
     implementation(project(':lib-streamsb-extractor'))
     implementation(project(':lib-dood-extractor'))
     implementation(project(':lib-voe-extractor'))
