+++ src/ar/anime4up/src/eu/kanade/tachiyomi/animeextension/ar/anime4up/extractors/StreamWishExtractor.kt
@@ -11,7 +11,7 @@ class StreamWishExtractor(private val client: OkHttpClient) {
     fun videosFromUrl(url: String, headers: Headers): List<Video> {
         val doc = client.newCall(GET(url)).execute().asJsoup()
         val script = doc.selectFirst("script:containsData(sources)")!!.data()
-        val scriptData = if(script.contains("eval")) JsUnpacker.unpackAndCombine(script)!! else script
+        val scriptData = if (script.contains("eval")) JsUnpacker.unpackAndCombine(script)!! else script
         val m3u8 = Regex("sources:\\s*\\[\\{\\s*\\t*file:\\s*[\"']([^\"']+)").find(scriptData)!!.groupValues[1]
         val streamLink = Regex("(.*)_,(.*),\\.urlset/master(.*)").find(m3u8)!!
         val streamQuality = streamLink.groupValues[2].split(",").reversed()
@@ -19,7 +19,7 @@ class StreamWishExtractor(private val client: OkHttpClient) {
         val qRegex = Regex("\".*?\":\\s*\"(.*?)\"").findAll(qualities)
         return qRegex.mapIndexed { index, matchResult ->
             val src = streamLink.groupValues[1] + "_" + streamQuality[index] + "/index-v1-a1" + streamLink.groupValues[3]
-            val quality = "Mirror: " +  matchResult.groupValues[1]
+            val quality = "Mirror: " + matchResult.groupValues[1]
             Video(src, quality, src, headers)
         }.toList()
     }
