+++ src/en/myanime/src/eu/kanade/tachiyomi/animeextension/en/myanime/Myanime.kt
@@ -5,7 +5,6 @@ import android.content.SharedPreferences
 import androidx.preference.ListPreference
 import androidx.preference.PreferenceScreen
 import eu.kanade.tachiyomi.animeextension.en.myanime.extractors.DailymotionExtractor
-import eu.kanade.tachiyomi.animeextension.en.myanime.extractors.GdrivePlayerExtractor
 import eu.kanade.tachiyomi.animeextension.en.myanime.extractors.YouTubeExtractor
 import eu.kanade.tachiyomi.animesource.ConfigurableAnimeSource
 import eu.kanade.tachiyomi.animesource.model.AnimeFilter
@@ -14,6 +13,7 @@ import eu.kanade.tachiyomi.animesource.model.SAnime
 import eu.kanade.tachiyomi.animesource.model.SEpisode
 import eu.kanade.tachiyomi.animesource.model.Video
 import eu.kanade.tachiyomi.animesource.online.ParsedAnimeHttpSource
+import eu.kanade.tachiyomi.lib.gdriveplayerextractor.GdrivePlayerExtractor
 import eu.kanade.tachiyomi.lib.okruextractor.OkruExtractor
 import eu.kanade.tachiyomi.network.GET
 import eu.kanade.tachiyomi.util.asJsoup
@@ -216,7 +216,8 @@ class Myanime : ConfigurableAnimeSource, ParsedAnimeHttpSource() {
                             YouTubeExtractor(client).videosFromUrl(url, "YouTube - ")
                         }
                         url.contains("gdriveplayer") -> {
-                            GdrivePlayerExtractor(client).videosFromUrl(url, name = "Gdriveplayer")
+                            val newHeaders = headersBuilder().add("Referer", baseUrl).build()
+                            GdrivePlayerExtractor(client).videosFromUrl(url, name = "Gdriveplayer", headers = newHeaders)
                         }
                         else -> null
                     }
