+++ src/id/neonime/build.gradle
@@ -11,6 +11,7 @@ ext {
 }
 
 dependencies {
+    implementation(project(':lib-gdriveplayer-extractor'))
     implementation(project(':lib-yourupload-extractor'))
     implementation(project(':lib-okru-extractor'))
     implementation "dev.datlag.jsunpacker:jsunpacker:1.0.1"
